requirements.txt
selene==2.0.0rc9
pytest
allure-pytest
selenium
python-dotenv==1.0.1